rename metadata editor
