+++ pwem/protocols/__init__.py
@@ -56,7 +56,6 @@ from .protocol_export import *
 from .protocol_tests import *
 
 from .protocol_origin_sampling_volume import ProtOrigSampling
-from .protocol_metadata_editor import ProtMetadataEditor
+from .protocol_set_editor import ProtSetEditor
 from .protocol_set_filter import ProtSetFilter
 from .protocol_movie_eraser import ProtMovieEraser
-
